+++ NEWS.md
@@ -1,12 +1,16 @@
 # bsvars 1.0.1.9000
 
 1.  Included Imports from package **stochvol**
-2.  Posterior computations for impulse responses and forecast error variance decomposition [#3](https://github.com/donotdespair/bsvars/issues/3)
+2.  Posterior computations for:
+
+-   impulse responses and forecast error variance decomposition [#3](https://github.com/donotdespair/bsvars/issues/3),
+-   structural shocks and historical decompositions [#14](https://github.com/donotdespair/bsvars/issues/14)
+-   fitted values [#17](https://github.com/donotdespair/bsvars/issues/17)
+
 3.  Implemented faster samplers based on random number generators from **armadillo** via **RcppArmadillo** [#7](https://github.com/donotdespair/bsvars/issues/7)
 4.  The `estimate_bsvar*` functions now also normalise the output w.r.t. to a structural matrix with positive elements on the main diagonal [#9](https://github.com/donotdespair/bsvars/issues/9)
 5.  Changed the order of arguments in the `estimate_bsvar*` functions with `posterior` first to facilitate workflows using the pipe `|>` [#10](https://github.com/donotdespair/bsvars/issues/10)
 6.  Include citation info for the package [#12](https://github.com/donotdespair/bsvars/issues/12)
-7.  Computations for structural shocks and historical decompositions [#14](https://github.com/donotdespair/bsvars/issues/14)
 
 # bsvars 1.0.0